Introduce shard splitting
Allow splitting a copy of a shard on a node.

HTTP API endepoint is `_shard_splits/*`

A post to `_shard_splits/jobs` with a node and a shard will create shard
splitting job. The job will copy all the data from source to the two targets,
then re-build indices, then switch the shard map to it and delete
(configurable) the source shard.

Most of the code is in the mem3_shard_split_* modules. The lower level logic to
perform the bulk copy of data is in the couch_db_split module.