Fix topoff
Previously we had a nasty bug where the internal replication filter was
crashing, but we caught the error and turned into `keep` response.

Instead of fixing that, it turns out we don't need that logic at and just use
the new feature of mem3_rep to replicate to multiple targets simoultaneously.